pyproject.toml
[project]
name = "ttp-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests",
    "beautifulsoup4",
    "rich",
    "mitreattack-python",
    "stix2",
    "python-dotenv",
    "OTXv2",
]
